Use separate file for MARGINS in UxTheme
